requirements.txt
streamlit>=1.28.0
groq>=0.4.0
openai>=1.0.0
streamlit-webrtc>=0.47.0
gtts>=2.3.0
numpy>=1.21.0
pydub>=0.25.0
requests>=2.28.0
sounddevice>=0.4.6
soundfile>=0.12.1
python-dotenv>=1.0.0
deepgram-sdk>=3.0.0

